webapi/tct-batterystatus-w3c-tests/batterystatus/w3c/battery-unplugging-dischargingTime-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: battery-unplugging-dischargingTime</title>
<meta name="flags" content="interact">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #note {
    background-color: #fef1b5;
    border: solid 1px #cdab2d;
    padding: 5px;
    margin: 15px;
    display: block;
  }
</style>

<h2>Description</h2>
<p>
  This test validates that ondischargingtimechange and dischargingTime attributes exist and
  are set to correct values, with corresponding events fired,
  when the charger is unplugged.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    The device is plugged in to the charger before this test is run.
  </li>
</ol>

<div id="note">
  Unplug the charger and wait for the test to complete.
</div>

<div id="log"></div>

<script>

(function() {

  setup({ explicit_timeout: true });

  var ondischargingtimechange_test = async_test(
      'When the device is unplugged in and its discharging time is updated, ' +
      'must set the dischargingTime attribute\'s value and ' +
      'fire a dischargingtimechange event.');

  var batterySuccess = function (battery) {
    battery.ondischargingtimechange = ondischargingtimechange_test.step_func(function () {
      assert_less_than(battery.dischargingTime, Infinity,
          'The value of the dischargingTime attribute must be set to the remaining discharging time.');
      ondischargingtimechange_test.done();
    });
  };

  var batteryFailure = function (error) {
    ondischargingtimechange_test.step(function () {
      assert_unreached(error.message);
    });
    ondischargingtimechange_test.done();
  };

  navigator.getBattery().then(batterySuccess, batteryFailure);

})();

</script>
